Ck productionalize (#11)
* chore(): test case skeleton

* docs(README.md): installation and prerequisites section

* docs(README.md): usage section

* chore(): linting pt 1

* chore(): linting pt 2

* test(): util test cases wip

* test(): util testing, add testing

* test(test/key-path.js): updated key-path tests

* test(): test case updates

* feat(): encapsulate password validation in util

* feat(): validate pw on commit

* fix(add.js): revert to use buffer constructor

* fix(add.js): revert createPipe return causing connection not closing

* test(): commit.js test cases

* test(): storage test cases

* chore(): lint new test files

* fix(add.js): add mirror dependency

* Update README.md

* docs(): README updates and ignored build dir

* docs(README.md): update general create cmd

+++ remove.js
@@ -1,24 +1,33 @@
-const debug = require('debug')('ara-filesystem:remove')
+/* eslint-disable no-await-in-loop */
+
 const { create } = require('./create')
-const pify = require('pify')
 
 async function remove({
   paths = [],
   did = '',
   password = ''
 } = {}) {
+  if (0 === paths.length) {
+    throw new Error('No path(s) provided.')
+  }
+
   const { afs } = await create({ did, password })
   for (const path of paths) {
+    if ('string' !== typeof path) {
+      throw new Error('Path found that is not of type string', path)
+    }
     try {
       if (await afs.access(path)) {
         await afs.unlink(path)
       }
-    } catch(err) {
+    } catch (err) {
       await afs.close()
-      return new Error("Could not remove file either because it does not exist or because of inadequate permissions")
+      return new Error(`Could not remove file either because it does not 
+        exist or because of inadequate permissions`)
     }
   }
   await afs.close()
+  return null
 }
 
 module.exports = {
